refactor: assigneeDto member 패키지 밑으로 이동하면서 리펙터링
