+++ src/main/java/dynamicquad/agilehub/issue/controller/response/IssueResponse.java
@@ -1,5 +1,6 @@
 package dynamicquad.agilehub.issue.controller.response;
 
+import dynamicquad.agilehub.member.dto.AssigneeDto;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -34,30 +35,6 @@ public class IssueResponse {
         public ContentDto() {
             this.text = "";
             this.imagesURLs = List.of();
-        }
-    }
-
-    @Builder
-    @Getter
-    @AllArgsConstructor
-    @EqualsAndHashCode
-    public static class AssigneeDto {
-        private Long id;
-        private String name;
-        private String profileImageURL;
-
-        public AssigneeDto() {
-            this.id = null;
-            this.name = "";
-            this.profileImageURL = "";
-        }
-
-        public static AssigneeDto from(Long id, String name, String profileImageURL) {
-            return AssigneeDto.builder()
-                .id(id)
-                .name(name)
-                .profileImageURL(profileImageURL)
-                .build();
         }
     }
