+++ src/main/java/dynamicquad/agilehub/issue/service/factory/IssueFactory.java
@@ -2,11 +2,11 @@ package dynamicquad.agilehub.issue.service.factory;
 
 import dynamicquad.agilehub.issue.controller.request.IssueRequest.IssueCreateRequest;
 import dynamicquad.agilehub.issue.controller.request.IssueRequest.IssueEditRequest;
-import dynamicquad.agilehub.issue.controller.response.IssueResponse.AssigneeDto;
 import dynamicquad.agilehub.issue.controller.response.IssueResponse.ContentDto;
 import dynamicquad.agilehub.issue.controller.response.IssueResponse.IssueDto;
 import dynamicquad.agilehub.issue.controller.response.IssueResponse.SubIssueDto;
 import dynamicquad.agilehub.issue.domain.Issue;
+import dynamicquad.agilehub.member.dto.AssigneeDto;
 import dynamicquad.agilehub.project.domain.Project;
 import java.util.List;
